Main script with everything
Added collision detection with motion clipping. Thank you Id software for the WOLF3D source code, I was stuck on this for a long while.
Added ability to switch to 2D map layout with a tab button press.
Blank blocks on the bmp map image are now supposed to be black in color instead of white.