Added SDL rendering
I've made a basic running application that can load a map into memory by reading a bitmap file on disk. This is very similar to what the pygame renderer does.
Moreover, there is also a simple game loop that keeps displaying the map onto the screen until the window is closed.