Added Hardware Rendering
I've used the SDL_Renderer to render everything on the screen, instead of the surface-based rendering that was used in all the previous versions of the engine. From now on I'll be working on this variant of the engine. I've also added an extra level on which I'm still working.